#607: Draft save/update routes
